adding HasImage as a concern
