+++ app/presenters/recipe_presenter.rb
@@ -1,4 +1,6 @@
 class RecipePresenter < Presenter
+  include HasImage
+
   alias_method :recipe, :element
 
   def self.model_name
@@ -7,27 +9,5 @@ class RecipePresenter < Presenter
 
   def ingredient_lines
     IngredientLinePresenter.wrap(recipe.ingredient_lines, recipe)
-  end
-
-  RANDOM_IMAGE_COUNT = 20
-  def image(size)
-    if recipe.image
-      recipe.image.url(size)
-    else
-      default_image(size)
-    end
-  end
-
-  def default_image(size)
-    hex = Digest::SHA1.hexdigest(recipe.title).to_i(16)
-    number = case size.to_s
-    when 'large'
-      hex % 2 == 0 ? 8 : 21
-    when 'full'
-      1
-    else
-      (hex % RANDOM_IMAGE_COUNT) + 1
-    end
-    "defaults/food/#{size}/#{number}.jpg"
   end
 end
